requirements.txt
font-fredoka-one==0.0.4
font-hanken-grotesk==0.0.2
font-intuitive==0.0.4
font-roboto==0.0.1
font-source-serif-pro==0.0.1
fonts==0.0.3
inky==1.2.2
numpy==1.21.4
Pillow==8.4.0
RPi.GPIO==0.7.0
smbus2==0.4.1
spidev==3.5
